refactor: ensure Model ItemType doesn't cause tsc OOM issue
Previously, tsc was running OOM when parsing mapped item types like
ItemTypeFromSchema. After thorough investigation, this issue was found
to be largely attributable to the ItemType type param of the Model class
which extends BaseItem, the values of which were set to the
SupportedAttributeValueTypes union. This meant that TS had to cache
each possible type / combination of types for recursively nested data,
which when fed into AscertainItemProcessingReturnType (which itself
accepts ItemType, Partial ItemType, and the array forms of each, and
then determines the corresponding type with an equally large set of
possible permutations), caused the tsc type-checker to run OOM.
This has been fixed by replacing the types-union in BaseItem with
'unknown', and replacing AscertainItemProcessingReturnType with
a more flexible generic which decentralizes the return-type logic by
allowing calling methods to provide type params that indicate the return
type.

As part of the implemented changes, several Model methods have also been
renamed/re-organized. Since the Model API has been changed, this commit
reflects a breaking change and should result in a major semver bump.

BREAKING CHANGE: Model API has been changed to fix tsc OOM issue.